deduplicate the time gathering code in Calendar
